Fine-tuning payment processing

+++ app/controllers/user_sessions_controller.rb
@@ -19,7 +19,7 @@ class UserSessionsController < ApplicationController
   end
 
   def destroy
-    current_user_session.destroy
+    current_user_session.destroy if current_user_session
     flash[:notice] = "You are now logged out."
     redirect_to home_path
   end
